Loadbalance control plane
Currently all incoming connections go to the first control-plane node.
This means that if kured is restarting that node, an API user can
experience downtime. This change loadbalances the incoming requests via
a loadbalancer, meaning the service never appears down.

To facilitate that, we also add a few new labels to resources to allow
for easy identification.

This also means the cluster will be highly available, as we can use the
loadbalancer rather than node IPs for the server ip address.