refactor: moving the shell scripting of the server installation outside of the hcloud_server block.
All the remote-exec blocks are now in a null-resource, instead of immediately in the hcloud_server block. The benefit is that now the private ip can be assigned (in the subnet) before microos is installed. With the current code base that has zero impact, but it opens up the possibility to ho to communicate safely in the private network at the installation stage already.